Fix phpstan reported doc block and casting issues.

+++ src/Http/Session.php
@@ -228,7 +228,7 @@ class Session
             $this->engine($class, $config['handler']);
         }
 
-        $this->_lifetime = ini_get('session.gc_maxlifetime');
+        $this->_lifetime = (int)ini_get('session.gc_maxlifetime');
         $this->_isCLI = (PHP_SAPI === 'cli' || PHP_SAPI === 'phpdbg');
         session_register_shutdown();
     }
